+++ tests/test_polynomial.py
@@ -7,9 +7,17 @@ Notes
   tested in separate testing modules.
 - Lagrange polynomials cannot be evaluated directly on a query point.
 """
+import itertools
 import numpy as np
 import pytest
-from conftest import assert_call, num_polynomials, polynomial_class
+
+from conftest import (
+    assert_call,
+    num_polynomials,
+    polynomial_class,
+    create_mi_pair_distinct,
+    POLY_CLASSES,
+)
 
 from minterpy import (
     LagrangePolynomial,
@@ -95,6 +103,168 @@ class TestInitialization:
         # Assertion
         with pytest.raises(ValueError):
             polynomial_class(mi)
+
+
+class TestEquality:
+    """All tests related to the equality check between polynomial instances."""
+
+    def test_single(
+        self,
+        polynomial_class,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+    ):
+        """Test equality between two instances with one set of coefficients."""
+        # Create a common multi-index set
+        mi = MultiIndexSet.from_degree(SpatialDimension, PolyDegree, LpDegree)
+
+        # Generate random coefficient values
+        coeffs = np.random.rand(len(mi))
+
+        # Create two equal polynomials
+        poly_1 = polynomial_class(mi, coeffs)
+        poly_2 = polynomial_class(mi, coeffs)
+
+        # Assertions
+        assert poly_1 is not poly_2  # Not identical instances
+        assert poly_1 == poly_2  # But equal in values
+        assert poly_2 == poly_1  # Symmetric property
+
+    def test_multiple(
+        self,
+        polynomial_class,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+        num_polynomials,
+    ):
+        """Test equality between two instances with multiple sets of coeffs."""
+        # Create a common multi-index set
+        mi = MultiIndexSet.from_degree(SpatialDimension, PolyDegree, LpDegree)\
+
+        # Generate random coefficient values
+        coeffs = np.random.rand(len(mi), num_polynomials)
+
+        # Create two equal polynomials
+        poly_1 = polynomial_class(mi, coeffs)
+        poly_2 = polynomial_class(mi, coeffs)
+
+        # Assertions
+        assert poly_1 is not poly_2  # Not identical instances
+        assert poly_1 == poly_2  # But equal in values
+        assert poly_2 == poly_1  # Symmetric property
+
+
+class TestInequality:
+    """All tests related to the inequality check of two poly. instances."""
+
+    def test_inconsistent_types(self, polynomial_class):
+        """Test inequality due to inconsistent type in the comparison."""
+        # Create a MultiIndexSet instance
+        mi = MultiIndexSet.from_degree(
+            spatial_dimension=3,
+            poly_degree=4,
+            lp_degree=2.0,
+        )
+
+        # Generate random coefficient values
+        coeffs = np.random.rand(len(mi))
+
+        # Create a polynomial instance
+        poly = polynomial_class(mi, coeffs)
+
+        # Assertions
+        assert poly != mi
+        assert poly != coeffs
+        assert poly != "123"
+        assert poly != 1
+        assert poly != 10.0
+
+    def test_different_multi_index(self, polynomial_class):
+        """Test inequality due to different multi-index sets."""
+        # Create two different multi-index sets
+        mi_1, mi_2 = create_mi_pair_distinct()
+
+        # Generate two sets of coefficients
+        coeffs_1 = np.random.rand(len(mi_1))
+        coeffs_2 = np.random.rand(len(mi_2))
+
+        # Create two polynomials
+        poly_1 = polynomial_class(mi_1, coeffs_1)
+        poly_2 = polynomial_class(mi_2, coeffs_2)
+
+        # Assertions
+        assert poly_1 != poly_2
+        assert poly_2 != poly_1
+
+    def test_different_grids(
+        self,
+        polynomial_class,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+    ):
+        """Test inequality due to different underlying Grids."""
+        # Create a common MultiIndexSet
+        mi = MultiIndexSet.from_degree(SpatialDimension, PolyDegree, LpDegree)
+
+        # Generate two grids
+        grd_1 = Grid(mi)
+        grd_2 = Grid.from_value_set(mi, np.linspace(-1, 1, PolyDegree+1))
+
+        # Generate a set of random coefficients
+        coeffs = np.random.rand(len(mi))
+
+        # Create two polynomials
+        poly_1 = polynomial_class(mi, coeffs, grid=grd_1)
+        poly_2 = polynomial_class(mi, coeffs, grid=grd_2)
+
+        # Assertions
+        assert poly_1 != poly_2
+        assert poly_2 != poly_1
+
+    def test_different_coeffs(
+        self,
+        polynomial_class,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+    ):
+        """Test inequality due to different coefficient values."""
+        # Create a common MultiIndexSet
+        mi = MultiIndexSet.from_degree(SpatialDimension, PolyDegree, LpDegree)
+
+        # Generate two sets of random coefficients
+        coeffs_1 = np.random.rand(len(mi))
+        coeffs_2 = np.random.rand(len(mi))
+
+        # Create two polynomials
+        poly_1 = polynomial_class(mi, coeffs_1)
+        poly_2 = polynomial_class(mi, coeffs_2)
+
+        # Assertions
+        assert poly_1 != poly_2
+        assert poly_2 != poly_1
+
+    def test_different_poly(self, SpatialDimension, PolyDegree, LpDegree):
+        """Test inequality due to different concrete polynomial classes."""
+        # Create a common MultiIndexSet
+        mi = MultiIndexSet.from_degree(SpatialDimension, PolyDegree, LpDegree)
+
+        # Generate a set of random coefficients
+        coeffs = np.random.rand(len(mi))
+
+        # Create two polynomials of different classes
+        poly_combinations = itertools.combinations(POLY_CLASSES, r=2)
+
+        for poly_class_1, poly_class_2 in poly_combinations:
+            poly_1 = poly_class_1(mi, coeffs)
+            poly_2 = poly_class_2(mi, coeffs)
+
+            # Assertions
+            assert poly_1 != poly_2
+            assert poly_2 != poly_1
 
 
 class TestEvaluation: